archwize/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
huggingface_hub==0.29.2
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6